requirements.txt
celery==4.4.7
celery[redis]
requests
vine==1.3.0
redis
tld
bs4
dnspython
lxml
mmh3==3.0.0
urllib3==1.26.5
XlsxWriter==1.4.4
colorama==0.4.4
beautifulsoup4==4.10.0
lxml
python-masscan
python-nmap
scikit-build
cmake
opencv-python
dnspython==2.1.0

